requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
chess==1.10.0
click==8.1.7
colorama==0.4.6
filelock==3.12.3
Flask==2.3.3
fsspec==2024.9.0
gunicorn==21.2.0
huggingface-hub==0.30.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.3
numpy==1.24.1
opencv-python==4.10.0.84
packaging==25.0
pandas==2.1.0
Pillow==9.3.0
PyAutoGUI==0.9.54
pygame==2.5.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
pywin32==306
PyYAML==6.0.2
qrcode==8.0
regex==2024.11.6
requests==2.28.1
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.15.2
torch==2.1.2
torchvision==0.2.0
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.7.1
tzdata==2025.2
urllib3==1.26.13
Werkzeug==2.3.7
wikipedia==1.4.0
